+++ src/main/java/ru/innopolis/stc9/dao/interfaces/AttendanceDao.java
@@ -5,11 +5,11 @@ import ru.innopolis.stc9.pojo.Attendance;
 import java.util.List;
 
 public interface AttendanceDao {
-    boolean addLessonAttendance(long lessonId, long[] students);
+    boolean addLessonAttendance(int lessonId, int[] students);
 
-    int getNumberOfMissedLessons(long id);
+    int getNumberOfMissedLessons(int id);
 
     boolean updateAttendance(Attendance attendance);
 
-    List<Attendance> getLessonAttendance(long lessonId, long groupId);
+    List<Attendance> getLessonAttendance(int lessonId, int groupId);
 }
