+++ src/main/java/ru/innopolis/stc9/dao/interfaces/MarkDao.java
@@ -5,9 +5,9 @@ import ru.innopolis.stc9.pojo.Mark;
 import java.util.List;
 
 public interface MarkDao {
-    List<Mark> getMarksByLessonId(long id);
+    List<Mark> getMarksByLessonId(int id);
 
-    Mark getMarkById(long id);
+    Mark getMarkById(int id);
 
     boolean addMark(Mark mark);
