+++ src/main/java/ru/innopolis/stc9/service/implementation/HomeWorkServiceImpl.java
@@ -41,30 +41,30 @@ public class HomeWorkServiceImpl implements HomeWorkService {
     }
 
     @Override
-    public HomeWork findById(long id) {
+    public HomeWork findById(int id) {
         return homeWorkDao.findById(id);
     }
 
     @Override
-    public List<HomeWork> getHomeWorkListByLessonId(long lessonId) {
+    public List<HomeWork> getHomeWorkListByLessonId(int lessonId) {
         return homeWorkDao.getHomeWorkListByLessonId(lessonId);
     }
 
     @Override
-    public HomeWork findByStudentId(long studentId) {
+    public HomeWork findByStudentId(int studentId) {
         return homeWorkDao.findByStudentId(studentId);
     }
 
     @Override
-    public HomeWork findByLessonId(long lessonId) {
+    public HomeWork findByLessonId(int lessonId) {
         return homeWorkDao.findByLessonId(lessonId);
     }
 
     @Override
-    public String findHomeWorkByMarkId(long markId) {
+    public String findHomeWorkByMarkId(int markId) {
         Mark mark = markDao.getMarkById(markId);
-        long studentId = mark.getUserId();
-        long lessonId = mark.getLessonId();
+        int studentId = mark.getUserId();
+        int lessonId = mark.getLessonId();
         HomeWork homeWork = homeWorkDao.findHomeWorkByStudentIdAndLessonId(studentId, lessonId);
         return homeWork.getHomeWorkURL();
     }
@@ -80,7 +80,7 @@ public class HomeWorkServiceImpl implements HomeWorkService {
     }
 
     @Override
-    public boolean deleteHomeWork(long id) {
+    public boolean deleteHomeWork(int id) {
         return homeWorkDao.deleteHomeWork(id);
     }
 }
