+++ src/main/java/ru/innopolis/stc9/service/interfaces/LessonsService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface LessonsService {
     boolean addLesson(Lessons lesson);
 
-    boolean deleteLesson(long lessonId);
+    boolean deleteLesson(int lessonId);
 
     List<Lessons> findAllLessons();
